Transition to separate subcomponent
